Fix JSON parsing

+++ json/src/main/java/btrplace/json/model/constraint/CumulatedResourceCapacityConverter.java
@@ -53,7 +53,7 @@ public class CumulatedResourceCapacityConverter extends SatConstraintConverter<C
         c.put("id", getJSONId());
         c.put("nodes", nodesToJSON(o.getInvolvedNodes()));
         c.put("rcId", o.getResource());
-        c.put("amount", (long) o.getAmount());
+        c.put("amount", o.getAmount());
         c.put("continuous", o.isContinuous());
         return c;
     }
