+++ json/src/main/java/btrplace/json/model/constraint/SingleResourceCapacityConverter.java
@@ -53,7 +53,7 @@ public class SingleResourceCapacityConverter extends SatConstraintConverter<Sing
         c.put("id", getJSONId());
         c.put("nodes", nodesToJSON(o.getInvolvedNodes()));
         c.put("rcId", o.getResource());
-        c.put("amount", (long) o.getAmount());
+        c.put("amount", o.getAmount());
         c.put("continuous", o.isContinuous());
         return c;
     }
