+++ json/src/main/java/btrplace/json/model/constraint/SingleRunningCapacityConverter.java
@@ -52,7 +52,7 @@ public class SingleRunningCapacityConverter extends SatConstraintConverter<Singl
         JSONObject c = new JSONObject();
         c.put("id", getJSONId());
         c.put("nodes", nodesToJSON(o.getInvolvedNodes()));
-        c.put("amount", (long) o.getAmount());
+        c.put("amount", o.getAmount());
         c.put("continuous", o.isContinuous());
         return c;
     }
